Add testing framework and enhance package structure
Introduced Pest testing framework, including its configuration and example tests. Updated composer dependencies to include necessary dev tools and testing libraries. Made updates to existing interfaces, traits, and migration files to fix comments and add clarity to functionality. Improved database setup and test support with proper UUID handling and model creation in test cases.

+++ src/database/migrations/2025_4_20_000001_update_lara_files_to_v2_table.php
@@ -27,14 +27,40 @@ return new class extends Migration
             });
         });
 
+        if (DB::getDriverName() === 'sqlite') {
+            Schema::rename('lara_files', 'lara_files_old');
+
+            Schema::create('lara_files', function (Blueprint $table) {
+                $table->increments('id'); // or uuid or whatever new type
+                $table->uuid()->nullable();
+                $table->string('disk')->default('public')->comment('Disk Adapter must be defined in your config/filesystems.php');
+                $table->string('path')->nullable()->comment('Relative file path.');
+                $table->string('hash_name')->nullable()->comment('Hashed name of the file.');
+                $table->string('extension')->nullable()->comment('Extension of the file.');
+                $table->string('name')->nullable()->comment('Original name of the file.');
+                $table->string('type')->nullable()->comment('Category of file. I.e. avatar, thumbnail, documents, etc.');
+                $table->string('visibility')->default(config('lara-files.visibility'))->comment('Browser visibility of the file.');
+                $table->text('description')->nullable()->comment('Description of the file.');
+                $table->unsignedInteger('author_id')->nullable()->comment('Author of the file.');
+                $table->string('larafilesable_type')->nullable()->comment('Name of the belonging model.');
+                $table->integer('larafilesable_id')->nullable()->comment('Id of the belonging model.');
+                $table->timestamps();
+            });
+
+            DB::statement('INSERT INTO lara_files SELECT * FROM lara_files_old');
+            Schema::drop('lara_files_old');
+        } else {
+            DB::statement('ALTER TABLE lara_files MODIFY id BIGINT UNSIGNED NOT NULL;');
+            DB::statement('ALTER TABLE lara_files DROP PRIMARY KEY;');
+        }
+
         // If you have any reference to lara_files_table . id in other tables , update here any FK columns and copy matching UUIDs .
 
-        DB::statement('ALTER TABLE lara_files MODIFY id BIGINT UNSIGNED NOT NULL;');
-        DB::statement('ALTER TABLE lara_files DROP PRIMARY KEY;');
-
-        Schema::table('lara_files', function (Blueprint $table) {
-            $table->dropColumn('id');
-        });
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::table('lara_files', function (Blueprint $table) {
+                $table->dropColumn('id');
+            });
+        }
 
         Schema::table('lara_files', function (Blueprint $table) {
             $table->renameColumn('uuid', 'id');
